refactor(nitro,nuxt): extract `@nuxt/nitro-server` package
