+++ packages/nitro-server/.attw.json
@@ -1,0 +1,3 @@
+{
+  "ignoreRules": ["cjs-resolves-to-esm"]
+}
