+++ packages/nitro-server/.gitignore
@@ -1,0 +1,2 @@
+src/runtime/templates/error-500.ts
+src/runtime/templates/error-dev.ts
