+++ packages/nitro-server/build.config.ts
@@ -1,0 +1,21 @@
+import { defineBuildConfig } from 'unbuild'
+import { addRollupTimingsPlugin, stubOptions } from '../../debug/build-config'
+
+export default defineBuildConfig({
+  declaration: true,
+  entries: [
+    'src/index',
+    { input: 'src/runtime/', outDir: 'dist/runtime', format: 'esm' },
+  ],
+  stubOptions,
+  hooks: {
+    'rollup:options' (ctx, options) {
+      addRollupTimingsPlugin(options)
+    },
+  },
+  externals: [
+    '@nuxt/schema',
+    'nuxt',
+    'nitropack',
+  ],
+})
