+++ packages/nitro-server/package.json
@@ -1,0 +1,59 @@
+{
+  "name": "@nuxt/nitro-server",
+  "version": "4.1.3",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/nuxt/nuxt.git",
+    "directory": "packages/nitro-server"
+  },
+  "description": "Nitro server integration for Nuxt",
+  "homepage": "https://nuxt.com",
+  "license": "MIT",
+  "type": "module",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": "./dist/index.mjs"
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "prepack": "unbuild",
+    "test:attw": "attw --pack"
+  },
+  "dependencies": {
+    "@nuxt/devalue": "^2.0.2",
+    "@nuxt/kit": "workspace:*",
+    "@unhead/vue": "^2.0.14",
+    "@vue/shared": "^3.5.22",
+    "consola": "^3.4.2",
+    "defu": "^6.1.4",
+    "destr": "^2.0.5",
+    "devalue": "^5.3.2",
+    "errx": "^0.1.0",
+    "escape-string-regexp": "^5.0.0",
+    "exsolve": "^1.0.7",
+    "h3": "^1.15.4",
+    "impound": "^1.0.0",
+    "klona": "^2.0.6",
+    "mocked-exports": "^0.1.1",
+    "nitropack": "^2.12.7",
+    "pathe": "^2.0.3",
+    "pkg-types": "^2.3.0",
+    "radix3": "^1.1.2",
+    "std-env": "^3.9.0",
+    "ufo": "^1.6.1",
+    "unctx": "^2.4.1",
+    "unstorage": "^1.17.1",
+    "vue": "^3.5.22",
+    "vue-bundle-renderer": "^2.2.0",
+    "vue-devtools-stub": "^0.1.0"
+  },
+  "devDependencies": {
+    "@nuxt/schema": "workspace:*",
+    "nuxt": "workspace:*",
+    "unbuild": "3.6.1",
+    "vitest": "3.2.4"
+  },
+  "engines": {}
+}
