+++ packages/nitro-server/src/runtime/handlers/renderer.ts
@@ -12,8 +12,7 @@ import { propsToString, renderSSRHead } from '@unhead/vue/server'
 import type { HeadEntryOptions, Link, Script } from '@unhead/vue/types'
 import destr from 'destr'
 import { defineRenderHandler, getRouteRules, useNitroApp } from 'nitropack/runtime'
-
-import type { NuxtPayload, NuxtSSRContext } from 'nuxt/app'
+import type { NuxtPayload, NuxtRenderHTMLContext, NuxtSSRContext } from 'nuxt/app'
 
 import { getRenderer } from '../utils/renderer/build-files'
 import { payloadCache } from '../utils/cache'
@@ -43,15 +42,6 @@ globalThis.__publicAssetsURL = publicAssetsURL
 // Polyfill for unctx (https://github.com/unjs/unctx#native-async-context)
 if (process.env.NUXT_ASYNC_CONTEXT && !('AsyncLocalStorage' in globalThis)) {
   (globalThis as any).AsyncLocalStorage = AsyncLocalStorage
-}
-
-export interface NuxtRenderHTMLContext {
-  htmlAttrs: string[]
-  head: string[]
-  bodyAttrs: string[]
-  bodyPrepend: string[]
-  body: string[]
-  bodyAppend: string[]
 }
 
 export interface NuxtRenderResponse {
