+++ packages/nitro-server/src/runtime/middleware/no-ssr.ts
