+++ packages/nitro-server/src/runtime/utils/cache.ts
