+++ packages/nitro-server/src/runtime/utils/error.ts
