+++ packages/nitro-server/src/runtime/utils/paths.ts
