+++ packages/nitro-server/src/runtime/utils/renderer/app.ts
@@ -1,9 +1,8 @@
 import type { H3Event } from 'h3'
 import { useRuntimeConfig } from 'nitropack/runtime'
 import { createHead } from '@unhead/vue/server'
+import type { NuxtPayload, NuxtSSRContext } from 'nuxt/app'
 import { sharedPrerenderCache } from '../cache'
-import type { NuxtSSRContext } from '#app'
-import type { NuxtPayload } from '#app/nuxt'
 // @ts-expect-error virtual file
 import unheadOptions from '#internal/unhead-options.mjs'
