+++ packages/nitro-server/src/runtime/utils/renderer/payload.ts
