+++ packages/nuxt/build.config.ts
@@ -36,6 +36,5 @@ export default defineBuildConfig({
     'nuxt/schema',
     '@vue/shared',
     '@unhead/vue',
-    'nitropack',
   ],
 })
