+++ packages/nuxt/package.json
@@ -70,9 +70,9 @@
   "dependencies": {
     "@dxup/nuxt": "^0.1.0",
     "@nuxt/cli": "^3.29.3",
-    "@nuxt/devalue": "^2.0.2",
     "@nuxt/devtools": "^2.6.5",
     "@nuxt/kit": "workspace:*",
+    "@nuxt/nitro-server": "workspace:*",
     "@nuxt/schema": "workspace:*",
     "@nuxt/telemetry": "^2.6.6",
     "@nuxt/vite-builder": "workspace:*",
@@ -87,9 +87,7 @@
     "destr": "^2.0.5",
     "devalue": "^5.3.2",
     "errx": "^0.1.0",
-    "esbuild": "^0.25.10",
     "escape-string-regexp": "^5.0.0",
-    "estree-walker": "^3.0.3",
     "exsolve": "^1.0.7",
     "h3": "^1.15.4",
     "hookable": "^5.5.3",
@@ -100,9 +98,7 @@
     "knitwork": "^1.2.0",
     "magic-string": "^0.30.19",
     "mlly": "^1.8.0",
-    "mocked-exports": "^0.1.1",
     "nanotar": "^0.2.0",
-    "nitropack": "^2.12.7",
     "nypm": "^0.6.2",
     "ofetch": "^1.4.1",
     "ohash": "^2.0.11",
@@ -117,7 +113,6 @@
     "radix3": "^1.1.2",
     "scule": "^1.3.0",
     "semver": "^7.7.3",
-    "seroval": "^1.3.2",
     "std-env": "^3.9.0",
     "tinyglobby": "^0.2.15",
     "ufo": "^1.6.1",
@@ -127,11 +122,8 @@
     "unimport": "^5.4.1",
     "unplugin": "^2.3.10",
     "unplugin-vue-router": "^0.15.0",
-    "unstorage": "^1.17.1",
     "untyped": "^2.0.0",
     "vue": "^3.5.22",
-    "vue-bundle-renderer": "^2.2.0",
-    "vue-devtools-stub": "^0.1.0",
     "vue-router": "^4.5.1"
   },
   "devDependencies": {
@@ -144,6 +136,7 @@
     "unbuild": "3.6.1",
     "vite": "7.1.9",
     "vitest": "3.2.4",
+    "vue-bundle-renderer": "2.2.0",
     "vue-sfc-transformer": "0.1.17"
   },
   "peerDependencies": {
