+++ packages/nuxt/types.d.mts
@@ -1,9 +1,4 @@
-/// <reference types="nitropack/types" />
-
-import type { H3Event } from 'h3'
-import type { LogObject } from 'consola'
-import type { NuxtIslandContext, NuxtIslandResponse, NuxtRenderHTMLContext } from './dist/app/types.js'
-import type { RuntimeConfig, SchemaDefinition } from 'nuxt/schema'
+import type { SchemaDefinition } from 'nuxt/schema'
 import type { DefineNuxtConfig } from 'nuxt/config'
 
 export * from './dist/index.js'
@@ -15,37 +10,5 @@ declare global {
   interface ImportMeta {
     url: string
     readonly env: ImportMetaEnv
-  }
-}
-
-// Note: Keep in sync with packages/nuxt/src/core/templates.ts
-declare module 'nitropack' {
-  // eslint-disable-next-line @typescript-eslint/no-empty-object-type
-  interface NitroRuntimeConfig extends RuntimeConfig {}
-  interface NitroRouteConfig {
-    ssr?: boolean
-    noScripts?: boolean
-    /** @deprecated Use `noScripts` instead */
-    experimentalNoScripts?: boolean
-  }
-  interface NitroRuntimeHooks {
-    'dev:ssr-logs': (ctx: { logs: LogObject[], path: string }) => void | Promise<void>
-    'render:html': (htmlContext: NuxtRenderHTMLContext, context: { event: H3Event }) => void | Promise<void>
-    'render:island': (islandResponse: NuxtIslandResponse, context: { event: H3Event, islandContext: NuxtIslandContext }) => void | Promise<void>
-  }
-}
-declare module 'nitropack/types' {
-  // eslint-disable-next-line @typescript-eslint/no-empty-object-type
-  interface NitroRuntimeConfig extends RuntimeConfig {}
-  interface NitroRouteConfig {
-    ssr?: boolean
-    noScripts?: boolean
-    /** @deprecated Use `noScripts` instead */
-    experimentalNoScripts?: boolean
-  }
-  interface NitroRuntimeHooks {
-    'dev:ssr-logs': (ctx: { logs: LogObject[], path: string }) => void | Promise<void>
-    'render:html': (htmlContext: NuxtRenderHTMLContext, context: { event: H3Event }) => void | Promise<void>
-    'render:island': (islandResponse: NuxtIslandResponse, context: { event: H3Event, islandContext: NuxtIslandContext }) => void | Promise<void>
   }
 }
