+++ packages/schema/src/types/schema.ts
@@ -1527,6 +1527,13 @@ export interface ConfigSchema {
   _modules: Array<any>
 
   /**
+   * Configuration for Nuxt's server builder.
+   */
+  server: {
+    builder?: '@nuxt/nitro-server' | (string & {}) | { bundle: (nuxt: Nuxt) => Promise<void> }
+  }
+
+  /**
    * Configuration for Nitro.
    *
    * @see [Nitro configuration docs](https://nitro.build/config)
